+++ packages/nativescript-masonkit/src-native/mason-android/masonkit/src/main/java/org/nativescript/mason/masonkit/LengthPercentageAuto.kt
@@ -10,17 +10,17 @@ sealed class LengthPercentageAuto {
 
   companion object {
     @JvmStatic
-    fun fromTypeValue(type: Int, value: Float): LengthPercentageAuto? {
+    fun fromTypeValue(type: Byte, value: Float): LengthPercentageAuto? {
       return when (type) {
-        0 -> Auto
-        1 -> Points(value)
-        2 -> Percent(value)
+        0.toByte() -> Auto
+        1.toByte() -> Points(value)
+        2.toByte() -> Percent(value)
         else -> null
       }
     }
   }
 
-  internal val type: Int
+  internal val type: Byte
     get() = when (this) {
       is Auto -> 0
       is Points -> 1
